renames tutorial to basics
