+++ basics/docs/CONFIGURATOR.md
@@ -6,7 +6,7 @@ The application we are building looks like this:
 
 ![Configurator sample](./configurator.gif)
 
-The code for the configurator example is located at `src/examples/configurator/index.js`.
+The code for the configurator example is located at `src/pages/configurator/index.js`.
 
 The root component looks like this:
 
@@ -32,8 +32,7 @@ class Configurator extends React.Component {
           <div>
             <p>To start the configurator specify a product id in the URL. </p>
             <p>
-              For examples{' '}
-              <a href="/configurator/IHEAR">/configurator/IHEAR</a>
+              For examples <a href="/configurator/IHEAR">/configurator/IHEAR</a>
             </p>
           </div>
         </Example>
@@ -80,54 +79,53 @@ class Configurator extends React.Component {
 Here, the `configure` function calls the API and updates the local state with the returned data:
 
 ```jsx
-  configure = async assignments => {
-    const { productId } = this.props.match.params;
-    const packagePath = process.env.REACT_APP_PACKAGE_PATH;
+configure = async assignments => {
+  const { productId } = this.props.match.params;
+  const packagePath = process.env.REACT_APP_PACKAGE_PATH;
 
-    try {
-      const result = await configure({
-        packagePath,
-        date: new Date(),
-        line: {
-          productId,
-          variableAssignments: assignments
-        }
-      });
-
-      this.setState({ sections: result.sections, error: null });
-    } catch (e) {
-      if (e.type === 'CannotLoadPackage') {
-        this.setState({
-          error: `Product with id '${productId} doesn't exist in package with path '${packagePath}'`
-        });
+  try {
+    const result = await configure({
+      packagePath,
+      date: new Date(),
+      line: {
+        productId,
+        variableAssignments: assignments
       }
+    });
+
+    this.setState({ sections: result.sections, error: null });
+  } catch (e) {
+    if (e.type === 'CannotLoadPackage') {
+      this.setState({
+        error: `Product with id '${productId} doesn't exist in package with path '${packagePath}'`
+      });
     }
-  };
+  }
+};
 ```
 
 In addition to this, we define three functions to handle the interaction when the user changes tab, assigns a value to a variable, and removes an assignment:
 
 ```jsx
-  handleActiveTabChange = activeTabIndex => this.setState({ activeTabIndex });
+handleActiveTabChange = activeTabIndex => this.setState({ activeTabIndex });
 
-  handleOnAssign = (variableId, value, exclude, multivalued) => {
-    const newAssignments = assign(
-      this.assignments,
-      { variableId, value, exclude },
-      multivalued
-    );
-    this.assignments = newAssignments;
-    this.configure(newAssignments);
-  };
+handleOnAssign = (variableId, value, exclude, multivalued) => {
+  const newAssignments = assign(
+    this.assignments,
+    { variableId, value, exclude },
+    multivalued
+  );
+  this.assignments = newAssignments;
+  this.configure(newAssignments);
+};
 
-  handleOnUnassign = (variableId, value, exclude) => {
-    const newAssignments = unassign(this.assignments, {
-      variableId,
-      value,
-      exclude
-    });
-    this.assignments = newAssignments;
-    this.configure(newAssignments);
-  };
+handleOnUnassign = (variableId, value, exclude) => {
+  const newAssignments = unassign(this.assignments, {
+    variableId,
+    value,
+    exclude
+  });
+  this.assignments = newAssignments;
+  this.configure(newAssignments);
+};
 ```
-
