+++ basics/docs/INTRO.md
@@ -1,0 +1,15 @@
+# Introduction
+
+The sample code illustrates how to use the CLM Platform configuration APIs.
+
+This introduction gives an overview of the code structure. The code this organized in the following folders:
+
+```
+├── api             # helpers for calling the http API
+├── components      # general components
+└── pages
+    ├── product-search  # example using the /products API
+    └── configurator    # example configurator using the /configure API
+```
+
+The entry to the examples is either the `pages/product-search/index.js` file or the `pages/configurator/index.js` file.
