+++ basics/docs/configurator.gif
