+++ basics/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "clm-sample-tutorials",
+  "name": "clm-basic-samples",
   "version": "1.0.0",
   "license": "MIT",
   "private": true,
