+++ basics/public/favicon.ico
