+++ basics/public/manifest.json
