+++ basics/src/pages/configurator/components/MultivaluedInput.css
