+++ basics/src/pages/configurator/components/Section.css
