+++ basics/src/pages/configurator/components/Section.js
