+++ basics/src/pages/configurator/components/Tabs.css
