+++ basics/src/pages/configurator/components/Tabs.js
