+++ basics/src/pages/configurator/components/TextInput.js
