+++ basics/src/pages/configurator/components/VariableInput.js
