+++ basics/src/pages/configurator/components/VariableLine.css
