+++ basics/src/pages/configurator/index.css
