+++ basics/src/pages/configurator/utils/assignment-utils.js
