+++ basics/src/pages/configurator/utils/variable-utils.js
