+++ basics/src/pages/product-search/index.css
