+++ basics/src/pages/product-search/index.js
