--- tutorial/docs/INTRO.md
@@ -1,12 +1,0 @@
-# Introduction
-
-The sample code illustrates how to use the CLM Platform configuration APIs.
-
-This introduction gives an overview of the code structure. The code this organized in the following folders:
-
-```
-+-- api             # helpers for calling the http API
-+-- components      # general components
-+-- product-search  # example for interacting with the /products API
-+-- configurator    # example configurator using the /configure API
-```
